requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]>=0.27.0
python-multipart==0.0.6
websockets==12.0

# AI Agents Framework
# NOTE: This project runs on Python 3.13. LangChain/LangGraph support modern Pydantic.
langchain>=0.2.16
langchain-core>=0.2.38
langchain-community>=0.2.16
langchain-openai>=0.1.25
langgraph>=0.2.30

# LLM Integration
openai==1.10.0
anthropic==0.8.1
tiktoken==0.5.2
google-generativeai==0.3.2
groq==0.4.1

# Web Automation
# NOTE: Optional (heavy / browser deps). Enable when using automation features.
# selenium==4.16.0
# playwright==1.41.0
# undetected-chromedriver==3.5.4
beautifulsoup4==4.12.3
lxml==5.1.0

# Vector Database & Embeddings
# NOTE: Optional (very heavy) and tends to cause long resolver/build times on Python 3.13.
# chromadb==0.4.22
# pinecone-client==6.0.0
# sentence-transformers==2.3.1
# faiss-cpu==1.9.0.post1

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
python-magic==0.4.27
pdfminer.six==20231228

# NLP & Text Processing (Optional - simplified for Python 3.13)
# spacy==3.7.2
nltk==3.8.1
scikit-learn>=1.6.0
# transformers==4.37.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
# psycopg2-binary fails to build on some Python 3.13 macOS setups without pg_config.
# psycopg provides prebuilt wheels and is the recommended modern driver.
psycopg[binary]==3.2.5
asyncpg==0.29.0
redis==5.0.1

# Data Validation & Serialization
pydantic>=2.7.4,<3.0.0
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
passlib==1.7.4
python-dotenv==1.0.0
cryptography>=40.0.0,<42.0.0
authlib==1.3.0

# Task Scheduling
celery==5.3.4
apscheduler==3.10.4

# Logging & Monitoring
loguru==0.7.2
sentry-sdk==1.40.0
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
aiohttp==3.9.1
tenacity==8.2.3

# Frontend (Streamlit)
# NOTE: Streamlit dashboard deps are optional and conflict with LangChain's packaging>=23.2 requirement
# (Streamlit 1.30 requires packaging<24). If/when you need the dashboard, install these separately.
# streamlit==1.30.0
# plotly==5.18.0
# Pillow==10.4.0
# pandas 2.1.x doesn't build cleanly on Python 3.13; use a newer wheel-enabled release.
pandas>=2.2.2

# Code Quality
black==23.12.1
pylint==3.0.3
flake8==7.0.0
mypy==1.8.0
bandit==1.7.6

# Email
sendgrid==6.11.0
email-validator==2.1.0

# Rate Limiting
slowapi==0.1.9

# CORS
# Handled via FastAPI/Starlette CORSMiddleware (no extra PyPI package needed)
